requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
motor==3.1.1
pydantic==1.10.12
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.6
email-validator==1.3.1
pymongo==4.3.3
bcrypt==4.0.1

